Add lesson41 task: Homework + tests

+++ src/test/java/gmail/taras/shevchuk0024/lesson37_Collections/NumberUtilsTest.java
@@ -14,7 +14,7 @@ class NumberUtilsTest {
         // given
         List<Integer> numbers = List.of(11, 22, 33, 44, 55, 66, 77, 88, 99);
         // when
-        int result = NumberUtils.getSum(numbers);
+        double result = NumberUtils.getSum(numbers);
         // then
         assertEquals(495, result);
     }
@@ -24,7 +24,7 @@ class NumberUtilsTest {
         // given
         List<Integer> numbers = List.of(11, 22, 33, 44, 55, 66, 77, 88, 99);
         // when
-        List<Integer> result = NumberUtils.multiplyOddNumber(numbers);
+        List<Long> result = NumberUtils.multiplyOddNumber(numbers);
         // then
         List<Integer> expected = List.of(22, 66, 110, 154, 198);
         assertEquals(expected, result);
